refactor(llrp): llrp is now an exported package
Allows the llrp package to be accessible for others to import.

Closes: #291
