+++ internal/driver/device.go
@@ -19,8 +19,8 @@ import (
 	"github.com/edgexfoundry/go-mod-core-contracts/v3/common"
 	contract "github.com/edgexfoundry/go-mod-core-contracts/v3/models"
 
-	"github.com/edgexfoundry/device-rfid-llrp-go/internal/llrp"
 	"github.com/edgexfoundry/device-rfid-llrp-go/internal/retry"
+	"github.com/edgexfoundry/device-rfid-llrp-go/pkg/llrp"
 )
 
 const (
