+++ internal/driver/discover.go
@@ -19,7 +19,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/edgexfoundry/device-rfid-llrp-go/internal/llrp"
+	"github.com/edgexfoundry/device-rfid-llrp-go/pkg/llrp"
 	dsModels "github.com/edgexfoundry/device-sdk-go/v3/pkg/models"
 	contract "github.com/edgexfoundry/go-mod-core-contracts/v3/models"
 )
