+++ pkg/llrp/binary_test.go
