+++ pkg/llrp/generated_structs.go
