+++ pkg/llrp/hacking_with_llrp.md
