+++ pkg/llrp/testdata/CloseConnectionResponse-000.bytes
