+++ pkg/llrp/testdata/GetAccessSpecsResponse-000.bytes
