+++ pkg/llrp/testdata/GetAccessSpecsResponse-000.json
