+++ pkg/llrp/testdata/GetROSpecsResponse-000.json
