+++ pkg/llrp/testdata/GetReaderCapabilitiesResponse-000.bytes
