+++ pkg/llrp/testdata/GetReaderConfigResponse-000.bytes
