+++ pkg/llrp/testdata/GetReaderConfigResponse-000.json
