+++ pkg/llrp/testdata/ReaderEventNotification-000.json
